Clean up superrask søknad code
