+++ src/modules/superrask-soknad/components/TrekkSoknad.js
@@ -1,0 +1,66 @@
+import React, { useEffect } from "react";
+import { FetchAction, FetchStatus, useFetchReducer } from "../../../common/hooks/useFetchReducer";
+import SearchAPI from "../../../common/api/SearchAPI";
+import { captureException } from "@sentry/browser";
+import DelayedSpinner from "../../../common/components/spinner/DelayedSpinner";
+import InterestAPI from "../api/InterestAPI";
+import TrekkSoknadSuccess from "./TrekkSoknadSuccess";
+import TrekkSoknadConfirmationRequired from "./TrekkSoknadConfirmationRequired";
+
+const TrekkSoknad = ({ match }) => {
+    const [{ data: ad, status: adFetchStatus }, dispatch] = useFetchReducer();
+    const [deleteSoknadResponse, deleteSoknadDispatch] = useFetchReducer();
+
+    /**
+     * Fetch ad
+     */
+    useEffect(() => {
+        const id = match.params.adUuid;
+
+        dispatch({ type: FetchAction.BEGIN });
+
+        SearchAPI.get(`api/stilling/${id}`)
+            .then((data) => {
+                dispatch({ type: FetchAction.RESOLVE, data });
+            })
+            .catch((error) => {
+                captureException(error);
+                dispatch({ type: FetchAction.REJECT, error });
+            });
+    }, []);
+
+    const handleWithDrawClick = () => {
+        deleteSoknadDispatch({ type: FetchAction.BEGIN });
+        InterestAPI.deleteInterest(match.params.adUuid, match.params.uuid)
+            .then((data) => {
+                deleteSoknadDispatch({ type: FetchAction.RESOLVE, data });
+            })
+            .catch((error) => {
+                captureException(error);
+                deleteSoknadDispatch({ type: FetchAction.REJECT, error });
+            });
+    };
+
+    return (
+        <div className="InterestMessageDelete">
+            {adFetchStatus === FetchStatus.IS_FETCHING && <DelayedSpinner />}
+
+            {(adFetchStatus === FetchStatus.SUCCESS || adFetchStatus === FetchStatus.FAILURE) && (
+                <React.Fragment>
+                    {deleteSoknadResponse.status !== FetchStatus.SUCCESS ? (
+                        <TrekkSoknadConfirmationRequired
+                            handleWithDrawClick={handleWithDrawClick}
+                            isDeleting={deleteSoknadResponse.status === FetchStatus.IS_FETCHING}
+                            hasError={deleteSoknadResponse.status === FetchStatus.FAILURE}
+                            ad={adFetchStatus === FetchStatus.SUCCESS ? ad : undefined}
+                        />
+                    ) : (
+                        <TrekkSoknadSuccess />
+                    )}
+                </React.Fragment>
+            )}
+        </div>
+    );
+};
+
+export default TrekkSoknad;
