+++ src/pages/superrask-soknad.js
@@ -1,6 +1,6 @@
 import React from "react";
 import useScrollToTop from "../common/hooks/useScrollToTop";
-import SuperraskSoknad from "../modules/superrask-soknad/SuperraskSoknad";
+import SuperraskSoknad from "../modules/superrask-soknad/components/SuperraskSoknad";
 import useDocumentTitle from "../common/hooks/useDocumentTitle";
 
 function SuperraskSoknadPage({ match }) {
